+++ Sources/SwiftUIKit/_Deprecated/Printing/StandardItemPrinter_iOS.swift
@@ -1,38 +1,17 @@
-//
-//  StandardItemPrinter.swift
-//  SwiftUIKit
-//
-//  Created by Daniel Saidi on 2022-04-04.
-//  Copyright © 2022 Daniel Saidi. All rights reserved.
-//
-
 #if os(iOS)
 import UIKit
 
-/**
- This class implements the ``ItemPrinter`` protocol by using
- a `UIPrintInteractionController` to print items.
- */
+@available(*, deprecated, message: "This has been moved to https://github.com/danielsaidi/printingkit.")
 public class StandardItemPrinter: ItemPrinter {
     
     public init() {}
     
-    /**
-     Whether or not the printer can print a certain item.
-     
-     - Parameters:
-       - item: The item to print.
-     */
     public func canPrint(_ item: PrintableItem) -> Bool {
-        return true
+        switch item {
+        case .pdf: return true
+        }
     }
     
-    /**
-     Print a list of printable items.
-     
-     - Parameters:
-       - item: The item to print.
-     */
     public func printItem(_ item: PrintableItem) {
         let printInfo = UIPrintInfo(dictionary: nil)
         printInfo.outputType = .general